Move to github workflows
This commit moves CI from circle to gh workflow. We also add a test for rails
main that is currently failing. I think its a binstub issue, and I'll look into this
later, but merging this anyway still seems valuable.

+++ superglue_rails/test/acceptance/superglue_installation_acceptance.rb
@@ -81,16 +81,6 @@ class SuperglueInstallationTest < Minitest::Test
     successfully %(npm pkg set scripts.build="#{build_script}")
   end
 
-  # def generate_test_app_6(app_name)
-  #   successfully "rails new #{app_name} \
-  #      --webpack \
-  #      --skip-git \
-  #      --skip-turbolinks \
-  #      --skip-hotwire \
-  #      --skip-spring \
-  #      --no-rc"
-  # end
-
   def generate_test_app_7(app_name)
     successfully "rails new #{app_name} \
        --javascript=esbuild \
@@ -164,7 +154,9 @@ class SuperglueInstallationTest < Minitest::Test
     click_button "Delete"
     assert page.has_content?("Post was successfully destroyed.")
   ensure
-    Process.kill "TERM", pid
-    Process.wait pid
+    if pid
+      Process.kill "TERM", pid
+      Process.wait pid
+    end
   end
 end
